Add files via upload
This is still work in progress, I have cleaned data and got a visuals for the price depending on Location, Number of rooms, Number of bathrooms. 
Duriing the regresion line it gets moving but nothing happens. 
I would need a bit of help with that althoguh I manage to get some visuals depending of times of purchase. 
Thanks
A